requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==4.4.2
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==73.0.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.17.8
wcwidth==0.2.13
Werkzeug==3.0.4
